+++ ActivEYE-Service/src/main/java/cz/muni/fi/pa165/activeye/service/RecordServiceImpl.java
@@ -2,6 +2,7 @@ package cz.muni.fi.pa165.activeye.service;
 
 import cz.muni.fi.pa165.activeye.dao.RecordDao;
 import cz.muni.fi.pa165.activeye.entities.Record;
+import cz.muni.fi.pa165.activeye.exceptions.ActiveyeDataAccessException;
 import org.springframework.stereotype.Service;
 
 import javax.inject.Inject;
@@ -9,6 +10,10 @@ import java.util.List;
 
 /**
  * Created by spriadka on 11/22/16.
+ */
+
+/**
+ * @author spriadka
  */
 @Service
 public class RecordServiceImpl implements RecordService {
@@ -19,21 +24,43 @@ public class RecordServiceImpl implements RecordService {
 
     @Override
     public void createRecord(Record record) {
-        recordDao.createRecord(record);
+        if (record == null){
+            throw new IllegalArgumentException("Record null");
+        }
+        try {
+            recordDao.createRecord(record);
+        }
+        catch (Exception e){
+            throw new ActiveyeDataAccessException("Error occured when creating record");
+        }
     }
 
     @Override
     public void updateRecord(Record record) {
-        recordDao.updateRecord(record);
+        if (record == null){
+            throw new IllegalArgumentException("Record null");
+        }
+        try {
+            recordDao.updateRecord(record);
+        }
+        catch(Exception e){
+            throw new ActiveyeDataAccessException("Error occured when updating record");
+        }
     }
 
     @Override
     public Record findById(Long id) {
+        if (id == null){
+            throw new IllegalArgumentException("Id cannot be null");
+        }
         return recordDao.getRecord(id);
     }
 
     @Override
     public void deleteRecord(Record record) {
+        if (record == null){
+            throw new IllegalArgumentException("Record cannot be null");
+        }
         recordDao.deleteRecord(record);
     }
